backend/requirements.txt
fastapi
uvicorn
groq
pydantic
python-dotenv
requests
python-multipart
pytesseract
pillow
psycopg2-binary
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
PyMuPDF
pdf2image
python-docx
pdfplumber
duckduckgo-search
reportlab
